refactor tests and move impl

+++ hypothesis-python/scripts/other-tests.sh
@@ -33,9 +33,11 @@ pip install "$(grep 'typing-extensions==' ../requirements/coverage.txt)"
 $PYTEST tests/typing_extensions/
 pip uninstall -y typing_extensions
 
-pip install "$(grep 'annotated-types==' ../requirements/coverage.txt)"
-$PYTEST tests/test_annotated_types.py
-pip uninstall -y annotated-types
+if [ "$(python -c 'import sys; print(sys.version_info[:2] >= (3, 9))')" = "True" ] ; then
+  pip install "$(grep 'annotated-types==' ../requirements/coverage.txt)"
+  $PYTEST tests/test_annotated_types.py
+  pip uninstall -y annotated-types
+fi
 
 pip install ".[lark]"
 pip install "$(grep -oE 'lark>=([0-9.]+)' ../hypothesis-python/setup.py | tr '>' =)"
